Improved referece rules.
Requiring a mutable Component reference in order to get a mutable trait
reference is too limiting and not neccesary so now you can get multiple
mutable trait refs as long as they are different traits while using an
immutable Component ref.

Though this isn't quite right: the check should be that the traits
are on different objects not that they are different traits.